third_party/blink/web_tests/webaudio/dom-exceptions-expected.txt

CONSOLE WARNING: The ScriptProcessorNode is deprecated. Use AudioWorkletNode instead. (https://bit.ly/audio-worklet)
CONSOLE WARNING: The ScriptProcessorNode is deprecated. Use AudioWorkletNode instead. (https://bit.ly/audio-worklet)
CONSOLE WARNING: The provided value 'fancy' is not a valid enum value of type ChannelCountMode.
CONSOLE WARNING: The provided value 'undefined' is not a valid enum value of type ChannelInterpretation.
CONSOLE WARNING: The provided value '9x' is not a valid enum value of type OverSampleType.
CONSOLE WARNING: The provided value 'junk' is not a valid enum value of type ChannelCountMode.
CONSOLE WARNING: The ScriptProcessorNode is deprecated. Use AudioWorkletNode instead. (https://bit.ly/audio-worklet)
CONSOLE WARNING: The provided value 'junk' is not a valid enum value of type ChannelCountMode.
This is a testharness.js-based test.
Harness: the test ran to completion.
